\name{carbonate}
\alias{carbonate}
\docType{data}
\title{
   ISE measurements of carbonate in seawater
}
\description{
A data set containing emf responses for 8 ISEs measuring carbonate in seawater
}
\usage{data(carbonate)}
\format{
	Load example carbonate data as an object of type ISEdata (see function loadISEdata)
}
\references{
  Dillingham, P.W., Alsaedi, B.S.O. and McGraw, C.M. (2017). Characterising uncertainty in instrumental limits of detection when sensor response is non-linear. \emph{2017 IEEE SENSORS}, Glasgow, United Kingdom, pp. 1-3. <doi:10.1109/ICSENS.2017.8233898>
}
\examples{
data(carbonate)
print(carbonate)
plot(carbonate)
}
\keyword{datasets}
