\name{print.ISEdescription}
\alias{print.ISEdescription}
\title{Prints tables of ISE parameters.}
\usage{
\method{print}{ISEdescription}(x, ...)
}
\arguments{
\item{x}{ISE analysis results (e.g. object of class analyseISE)}

\item{...}{Other objects passed through.}
}
\value{No return value, prints results from describeISE.}
\description{
Prints tables of ISE parameters for one or multiple ISEs.
}
\seealso{
\code{\link{describeISE}}
}
\author{
Peter Dillingham, \email{peter.dillingham@otago.ac.nz}
}
